requirements.txt
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.5.3
cryptography==42.0.8
decorator==5.1.1
Deprecated==1.2.14
fabric==3.2.2
idna==3.7
invoke==2.2.0
paramiko==3.4.0
pretty-json==1.2.0
pycparser==2.22
Pygments==2.18.0
pygments-json==0.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
urllib3==2.2.1
wrapt==1.16.0
